Задание 1/ocr_service/pyproject.toml
[project]
name = "ocr-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "surya-ocr>=0.17",
    "python-multipart>=0.0.9",
    "pillow>=10.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]
